backend/requirements.txt
fastapi
uvicorn
pandas
numpy
python-multipart
pydantic
python-dotenv
google-generativeai
scikit-learn
fastapi-cors
